Add first draft of abstraction for the ts3 layer.
